Fix import structure
