+++ optuna/integration/_allennlp/executor.py
@@ -1,7 +1,6 @@
 import json
 import os
 from typing import Any
-from typing import Callable
 from typing import Dict
 from typing import List
 from typing import Optional
@@ -10,6 +9,7 @@ from typing import Union
 import optuna
 from optuna._experimental import experimental
 from optuna._imports import try_import
+from optuna.integration._allennlp.environment import _environment_variables
 
 
 with try_import() as _imports:
@@ -22,29 +22,6 @@ with try_import() as _imports:
 # the environment that builds the documentation.
 if _imports.is_successful():
     import _jsonnet
-
-    from optuna.integration._allennlp.environment import _environment_variables
-else:
-    # I disable mypy here since `allennlp.training.TrainerCallback` is a subclass of `Registrable`
-    # (https://docs.allennlp.org/main/api/training/trainer/#trainercallback) but `TrainerCallback`
-    # defined here is not `Registrable`, which causes a mypy checking failure.
-    class TrainerCallback:  # type: ignore
-        """Stub for TrainerCallback."""
-
-        @classmethod
-        def register(cls: Any, *args: Any, **kwargs: Any) -> Callable:
-            """Stub method for `TrainerCallback.register`.
-
-            This method has the same signature as
-            `Registrable.register <https://docs.allennlp.org/master/
-            api/common/registrable/#registrable>`_ in AllenNLP.
-
-            """
-
-            def wrapper(subclass: Any, *args: Any, **kwargs: Any) -> Any:
-                return subclass
-
-            return wrapper
 
 
 _PPID = os.getppid()
